requirements.txt
numpy==1.13.3
librosa==0.8.1
pandas==1.1.5
tensorflow==2.6.0
tqdm==4.62.3
PyAudio==0.2.11
SpeechRecognition==3.8.1
PyYAML==5.4.1
scikit_learn==1.0
